Prepare CYF branch
  - Remove tests and CI
  - Update README
  - Add MongoDB
