+++ .cfignore
@@ -1,8 +1,6 @@
-__mocks__/
 .github/
 .idea/
 .vscode/
 client/
-e2e/
 node_modules/
 server/
