+++ .dockerignore
@@ -1,9 +1,7 @@
-__mocks__/
 .git/
 .github/
 .idea/
 .vscode/
 bin/
 dist/
-e2e/
 node_modules/
