+++ .eslintignore
@@ -1,3 +1,1 @@
-.stryker-tmp/
 dist/
-reports/
