+++ .gitignore
@@ -1,8 +1,4 @@
 .idea/
-.stryker-tmp/
 dist/
-e2e/screenshots/
-e2e/videos/
 node_modules/
-reports/
 .DS_Store
