+++ .slugignore
@@ -1,4 +1,3 @@
 /.github
 /.idea
 /.vscode
-/e2e
