--- .travis.yml
@@ -1,62 +1,0 @@
-_no-cypress: &no-cypress
-  addons: {}
-  env:
-    - CYPRESS_INSTALL_BINARY=0
-
-dist: bionic
-
-language: node_js
-
-node_js:
-  - 12.18.3
-  - 14
-  - 12
-  - 10
-
-addons:
-  apt:
-    packages:
-      - libgconf-2-4
-
-cache:
-  npm: true
-  directories:
-    - ~/.cache  # Cypress binary
-
-script:
-  - npm run e2e
-
-jobs:
-  include:
-    - stage: build
-      <<: *no-cypress
-      services:
-        - docker
-      before_script: ./bin/check-node.sh
-      script:
-        - npm run lint
-        - npm test
-        - docker build .
-    - stage: deploy
-      <<: *no-cypress
-      install: true
-      script: true
-      before_deploy: ./bin/prepare-deploy.sh
-      deploy:
-        - provider: cloudfoundry
-          skip_cleanup: true
-          api: https://api.run.pivotal.io
-          username: $CF_USERNAME
-          password: $CF_PASSWORD
-          organization: $CF_ORG
-          space: $CF_SPACE
-          on:
-            branch: master
-        - provider: heroku
-          api_key: $HEROKU_API_KEY
-          app: starter-kit-ci
-
-stages:
-  - build
-  - test
-  - deploy
