+++ Dockerfile
@@ -4,7 +4,6 @@ FROM node:${NODE_RELEASE}-alpine AS build
 
 ARG NODE_RELEASE
 
-ENV CYPRESS_INSTALL_BINARY=0
 
 RUN echo "Node $(node -v) / NPM v$(npm -v)"
