--- cypress.json
@@ -1,10 +1,0 @@
-{
-  "baseUrl": "http://localhost:3000",
-  "integrationFolder": "e2e/integration",
-  "fixturesFolder": "e2e/fixtures",
-  "pluginsFile": "e2e/plugins",
-  "screenshotsFolder": "e2e/screenshots",
-  "supportFile": "e2e/support",
-  "video": false,
-  "videosFolder": "e2e/videos"
-}
