--- e2e/.eslintrc.json
@@ -1,11 +1,0 @@
-{
-  "env": {
-    "cypress/globals": true
-  },
-  "parserOptions": {
-    "ecmaVersion": 6
-  },
-  "plugins": [
-    "cypress"
-  ]
-}
