--- e2e/integration/home.test.js
@@ -1,12 +1,0 @@
-beforeEach(() => {
-	cy.visit("/");
-});
-
-it("displays the site", () => {
-	cy.get("[data-qa=\"message\"]").should("contain.text", "Hello, world!");
-});
-
-it("meets basic accessibility guidelines", () => {
-	cy.injectAxe();
-	cy.checkA11y();
-});
