+++ package.json
@@ -6,26 +6,16 @@
   "scripts": {
     "prebuild": "npm run cleanup",
     "build": "npm run build:server && npm run build:client",
-    "postbuild": "./bin/build-info.sh || echo 'No build info'",
     "build:client": "webpack --config client/webpack/prod.config.js",
-    "build:server": "babel server --out-dir dist --ignore \"**/*.test.js,server/static\"",
+    "build:server": "babel server --out-dir dist",
     "cleanup": "rimraf ./dist/*",
     "dev": "concurrently -k -n \"client,server\" \"npm run dev:client\" \"npm run dev:server\"",
     "dev:client": "webpack-dev-server --config client/webpack/dev.config.js",
     "dev:server": "cross-env PORT=3100 nodemon --inspect --watch server --exec babel-node server/server.js",
-    "e2e": "concurrently -k -s first -n \"app,e2e\" \"npm run serve\" \"npm run e2e:safe\"",
-    "e2e:dev": "concurrently -k -s first -n \"dev,e2e\" \"npm run dev\" \"npm run e2e:safe\"",
-    "e2e:local": "cypress open",
-    "e2e:run": "cypress run",
-    "e2e:safe": "wait-on -l http-get://localhost:3000 && npm run e2e:run",
     "lint": "eslint .",
     "preserve": "npm run build",
     "serve": "npm start",
-    "ship": "npm run lint && npm run test && npm run e2e:dev && npm run e2e",
-    "start": "node dist/server.js",
-    "test": "jest",
-    "test:mutate": "stryker run",
-    "test:watch": "npm run test -- --watch"
+    "start": "node dist/server.js"
   },
   "repository": {
     "type": "git",
@@ -47,38 +37,25 @@
     "@babel/preset-env": "^7.10.4",
     "@babel/preset-react": "^7.10.4",
     "@codeyourfuture/eslint-config-standard": "^2.0.1",
-    "@stryker-mutator/core": "^3.3.1",
-    "@stryker-mutator/javascript-mutator": "^3.3.1",
-    "@stryker-mutator/jest-runner": "^3.3.1",
-    "@testing-library/jest-dom": "^5.11.2",
-    "@testing-library/react": "^10.4.7",
-    "axe-core": "^3.5.5",
     "axios": "^0.19.2",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.1.0",
     "concurrently": "^5.2.0",
     "cross-env": "^7.0.2",
     "css-loader": "^4.1.0",
-    "cypress": "^4.11.0",
-    "cypress-axe": "^0.8.1",
     "eslint": "^7.5.0",
-    "eslint-plugin-cypress": "^2.11.1",
-    "eslint-plugin-jest": "^23.19.0",
     "eslint-plugin-jsx-a11y": "^6.3.1",
     "eslint-plugin-react": "^7.20.5",
     "eslint-plugin-react-hooks": "^4.0.8",
     "file-loader": "^6.0.0",
     "html-webpack-plugin": "^4.3.0",
     "html-webpack-tags-plugin": "^2.0.17",
-    "jest": "^26.1.0",
-    "jest-circus": "^26.1.0",
     "nodemon": "^2.0.4",
     "prop-types": "^15.7.2",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "rimraf": "^3.0.2",
     "style-loader": "^1.2.1",
-    "supertest": "^4.0.2",
     "wait-on": "^5.1.0",
     "webpack": "^4.44.0",
     "webpack-cli": "^3.3.12",
@@ -89,6 +66,7 @@
     "@babel/runtime": "^7.10.5",
     "express": "^4.17.1",
     "helmet": "^3.23.3",
+    "mongodb": "^3.6.0",
     "morgan": "^1.10.0"
   },
   "engines": {
